Deprecate sync trtllm_worker code
