+++ nemo_rl/models/generation/trtllm/trtllm_worker_async.py
@@ -14,16 +14,19 @@
 
 """Ray actor wrapping ``tensorrt_llm._torch.async_llm.AsyncLLM``.
 
-Sibling of :class:`TrtllmGenerationWorker` for the async TRT-LLM engine.
-Every method that calls into ``AsyncLLM`` is exposed as ``async def`` with
-the ``_async`` suffix, so Ray's actor runtime runs them on the actor's own
-asyncio loop. Sync entrypoints inherited from
-:class:`TrtllmGenerationWorkerImpl` (e.g. ``shutdown``) stay sync.
-:class:`TrtllmGeneration` dispatches to the ``_async`` variants when
-``trtllm_cfg.async_engine`` is set.
+The sole TRT-LLM generation worker (the synchronous engine path was removed;
+see :class:`TrtllmGeneration`, which asserts ``trtllm_cfg.async_engine=true``).
+Every method that calls into ``AsyncLLM`` is exposed as ``async def`` with the
+``_async`` suffix, so Ray's actor runtime runs them on the actor's own asyncio
+loop; process-lifecycle / helper methods (e.g. ``shutdown``,
+``configure_worker``) stay sync.
+
+Weight updates flow through ``NcclExtension`` inside TRT-LLM's internal
+``RayGPUWorker``, invoked via ``llm.collective_rpc()``.
 """
 
 import asyncio
+import gc
 import os
 from typing import Any, Optional
 
@@ -31,6 +34,7 @@ import ray
 import torch
 
 from nemo_rl.distributed.batched_data_dict import BatchedDataDict
+from nemo_rl.distributed.virtual_cluster import PY_EXECUTABLES
 from nemo_rl.distributed.worker_group_utils import get_nsight_config_if_pattern_matches
 from nemo_rl.models.generation.interfaces import (
     GenerationDatumSpec,
@@ -38,11 +42,44 @@ from nemo_rl.models.generation.interfaces import (
     verify_right_padding,
 )
 from nemo_rl.models.generation.trtllm.config import TrtllmConfig
-from nemo_rl.models.generation.trtllm.trtllm_worker import TrtllmGenerationWorkerImpl
 
 
-class TrtllmAsyncGenerationWorkerImpl(TrtllmGenerationWorkerImpl):
-    """Async variant of :class:`TrtllmGenerationWorkerImpl` (plain class)."""
+class TrtllmAsyncGenerationWorkerImpl:
+    """Plain (non-actor) implementation of the async TRT-LLM generation worker.
+
+    Held separately from the ``@ray.remote``-wrapped
+    :class:`TrtllmAsyncGenerationWorker` so it can be exercised without Ray.
+    """
+
+    @staticmethod
+    def configure_worker(
+        num_gpus: int | float,
+        bundle_indices: Optional[tuple[int, list[int]]] = None,
+    ) -> tuple[dict[str, Any], dict[str, str], dict[str, Any], dict[str, Any]]:
+        # TRT-LLM with orchestrator_type="ray" creates its own internal
+        # Ray actors that each need a GPU.  The outer actor therefore
+        # gives up its GPU reservation.
+        resources: dict[str, Any] = {"num_gpus": 0, "num_cpus": 0}
+        # TRT-LLM's CudaRunner derives NVRTC -I include paths via
+        # `popen("pip show tensorrt_llm")`. Pin the worker's actor python's
+        # bin dir to the front of PATH so `pip` resolves to the interpreter
+        # whose site-packages contain tensorrt_llm.
+        worker_py_bin_dir = os.path.dirname(PY_EXECUTABLES.TRTLLM)
+        worker_path = f"{worker_py_bin_dir}:" + os.environ.get("PATH", "")
+        env_vars: dict[str, str] = {
+            "RAY_EXPERIMENTAL_NOSET_CUDA_VISIBLE_DEVICES": "1",
+            "NCCL_CUMEM_ENABLE": "1",
+            "PATH": worker_path,
+        }
+        init_kwargs: dict[str, Any] = {}
+
+        if bundle_indices is not None:
+            # Pass bundle_indices through __init__ kwargs; the worker resolves the
+            # parent placement group via get_current_placement_group() and hands both
+            # to TRT-LLM as ray_placement_config (instead of TRTLLM_RAY_BUNDLE_INDICES).
+            init_kwargs["bundle_indices"] = bundle_indices[1]
+
+        return resources, env_vars, init_kwargs, {}
 
     def __repr__(self) -> str:
         return "TrtllmAsyncGenerationWorker"
@@ -185,6 +222,9 @@ class TrtllmAsyncGenerationWorkerImpl(TrtllmGenerationWorkerImpl):
     #  Lifecycle
     # ------------------------------------------------------------------ #
 
+    def is_alive(self) -> bool:
+        return True
+
     async def post_init_async(self) -> None:
         """Finish async-side engine setup and (optionally) start HTTP server."""
         if not self.is_model_owner or self.llm is None:
@@ -199,7 +239,16 @@ class TrtllmAsyncGenerationWorkerImpl(TrtllmGenerationWorkerImpl):
 
     def shutdown(self) -> bool:
         self.stop_http_server()
-        return super().shutdown()
+        try:
+            if self.llm is not None:
+                del self.llm
+                self.llm = None
+            gc.collect()
+            torch.cuda.empty_cache()
+            return True
+        except Exception as e:
+            print(f"Error during TRT-LLM shutdown: {e}")
+            return False
 
     # ------------------------------------------------------------------ #
     #  HTTP server for NeMo Gym
@@ -352,7 +401,6 @@ class TrtllmAsyncGenerationWorkerImpl(TrtllmGenerationWorkerImpl):
             return True
         await self.reset_prefix_cache_async()
         await self.llm.release(self._all_sleep_tags())
-        import gc
         gc.collect()
         torch.cuda.empty_cache()
         return True
@@ -459,6 +507,52 @@ class TrtllmAsyncGenerationWorkerImpl(TrtllmGenerationWorkerImpl):
                 unpadded_sequence_lengths, dtype=torch.long
             ),
         })
+
+    # ------------------------------------------------------------------ #
+    #  Helpers
+    # ------------------------------------------------------------------ #
+
+    def _resolve_end_id(self) -> Optional[int]:
+        """Resolve end_id from model config.json, cached after first call.
+
+        Mirrors vLLM engine which reads eos_token_id from config.json automatically
+        at startup. TRT-LLM requires it to be passed explicitly as end_id.
+        """
+        if hasattr(self, "_end_id_cache"):
+            return self._end_id_cache
+        end_id: Optional[int] = None
+        try:
+            from transformers import AutoConfig
+            hf_config = AutoConfig.from_pretrained(self.model_name, trust_remote_code=True)
+            eos_id = getattr(hf_config, "eos_token_id", None)
+            if eos_id is not None:
+                end_id = eos_id[0] if isinstance(eos_id, list) else eos_id
+        except Exception as e:
+            print(f"[TrtllmAsyncWorker] AutoConfig load failed: {e}", flush=True)
+        self._end_id_cache = end_id
+        return end_id
+
+    def _build_sampling_params(self, *, greedy: bool):
+        top_k_cfg = self.cfg["top_k"]
+        top_k_val = 1 if greedy else (top_k_cfg if top_k_cfg is not None else 0)
+        temperature = 0.0 if greedy else self.cfg["temperature"]
+
+        end_id = self._resolve_end_id()
+        stop_ids = list(self.cfg.get("stop_token_ids") or [])
+
+        return self.TrtSamplingParams(
+            temperature=temperature,
+            top_p=self.cfg["top_p"],
+            top_k=top_k_val,
+            max_tokens=self.cfg["max_new_tokens"],
+            end_id=end_id,
+            stop_token_ids=stop_ids or None,
+            # Keep the EOS / stop token in the returned token_ids so that the
+            # response sequence matches HF / vLLM behavior. Required for
+            # logprob alignment with training-side Megatron.
+            include_stop_str_in_output=True,
+            logprobs=True,
+        )
 
 
 @ray.remote(
